Create MessageUtils
